Implement tests for new DevQL stages: coverage, deps, and tests
This commit introduces a suite of tests for the newly added DevQL stages, including coverage, deps, and tests. The tests validate the correct parsing and execution of queries involving these stages, ensuring that they adhere to the expected behavior and constraints. Additionally, it includes checks for stage combinations that are not allowed, enhancing the overall robustness of the query pipeline. New files are created for each stage's tests, contributing to improved test coverage and reliability.
