+++ bitloops/src/host/devql/tests/devql_tests/query_pipeline/parsing.rs
@@ -1,0 +1,58 @@
+use super::super::*;
+
+#[test]
+fn parse_devql_pipeline_basic() {
+    let parsed = parse_devql_query(
+        r#"repo("bitloops-cli")->asOf(ref:"main")->file("src/main.rs")->artefacts(lines:1..50,kind:"file",agent:"claude-code",since:"2026-03-01")->select(path,canonical_kind)->limit(10)"#,
+    )
+    .unwrap();
+
+    assert_eq!(parsed.repo.as_deref(), Some("bitloops-cli"));
+    assert!(matches!(parsed.as_of, Some(AsOfSelector::Ref(ref v)) if v == "main"));
+    assert_eq!(parsed.file.as_deref(), Some("src/main.rs"));
+    assert_eq!(parsed.artefacts.kind.as_deref(), Some("file"));
+    assert_eq!(parsed.artefacts.lines, Some((1, 50)));
+    assert_eq!(parsed.artefacts.agent.as_deref(), Some("claude-code"));
+    assert_eq!(parsed.artefacts.since.as_deref(), Some("2026-03-01"));
+    assert_eq!(parsed.limit, 10);
+    assert_eq!(parsed.select_fields, vec!["path", "canonical_kind"]);
+}
+
+#[test]
+fn parse_devql_artefacts_symbol_fqn_filter() {
+    let parsed = parse_devql_query(
+        r#"repo("rust-example")->artefacts(kind:"method",symbol_fqn:"hello_rust/src/main.rs::impl@1::handle_factorial")->limit(5)"#,
+    )
+    .unwrap();
+
+    assert_eq!(parsed.artefacts.kind.as_deref(), Some("method"));
+    assert_eq!(
+        parsed.artefacts.symbol_fqn.as_deref(),
+        Some("hello_rust/src/main.rs::impl@1::handle_factorial")
+    );
+}
+
+#[test]
+fn parse_devql_checkpoints_basic() {
+    let parsed = parse_devql_query(
+        r#"repo("bitloops-cli")->checkpoints(agent:"claude-code",since:"2026-03-01")->select(checkpoint_id,created_at)->limit(5)"#,
+    )
+    .unwrap();
+
+    assert!(parsed.has_checkpoints_stage);
+    assert_eq!(parsed.checkpoints.agent.as_deref(), Some("claude-code"));
+    assert_eq!(parsed.checkpoints.since.as_deref(), Some("2026-03-01"));
+    assert_eq!(parsed.limit, 5);
+}
+
+#[test]
+fn parse_devql_chat_history_stage_basic() {
+    let parsed = parse_devql_query(
+        r#"repo("bitloops-cli")->file("index.ts")->artefacts(lines:1..10)->chatHistory()->limit(3)"#,
+    )
+    .unwrap();
+
+    assert!(parsed.has_artefacts_stage);
+    assert!(parsed.has_chat_history_stage);
+    assert_eq!(parsed.limit, 3);
+}
