+++ bitloops/src/host/devql/tests/devql_tests/query_pipeline/pipeline_validation.rs
@@ -1,0 +1,31 @@
+use super::super::*;
+
+#[tokio::test]
+async fn execute_devql_query_rejects_chat_history_without_artefacts_stage() {
+    let cfg = test_cfg();
+    let events_cfg = default_events_cfg();
+    let parsed = parse_devql_query(r#"repo("temp2")->chatHistory()->limit(1)"#).unwrap();
+    let err = execute_devql_query(&cfg, &parsed, &events_cfg, None)
+        .await
+        .unwrap_err();
+    assert!(
+        err.to_string()
+            .contains("chatHistory() requires an artefacts() stage")
+    );
+}
+
+#[tokio::test]
+async fn execute_devql_query_rejects_combining_checkpoints_and_artefacts_stage() {
+    let cfg = test_cfg();
+    let events_cfg = default_events_cfg();
+    let parsed =
+        parse_devql_query(r#"repo("temp2")->checkpoints()->artefacts(agent:"claude-code")"#)
+            .unwrap();
+    let err = execute_devql_query(&cfg, &parsed, &events_cfg, None)
+        .await
+        .unwrap_err();
+    assert!(
+        err.to_string()
+            .contains("MVP limitation: telemetry/checkpoints stages cannot be combined")
+    );
+}
